Add files via upload
This was my HTML and CSS page that I created. It was a side project that provided links to popular Cisco Jabber (a software phone) user guides. The user would click on the appropriate icon and then would be led to the respective Jabber user guide on the Pfizer BT on Demand (tech support website). 

Note: for data sensitivity purposes, the links will not work unless your device is a registered Pfizer device on their network.